Exploiting CVE-2013-7285 (XStream) - tab 12: Update Vulnerable Component
Webgoat originally used com.thoughtworks.xstream v.1.4.7, which
is prone to XXE. Updated to the latest version (1.4.17) according
to the Xstream website (http://x-stream.github.io/changes.html).
Fix also configured xstream object to just accept input type of
Contact class.

During this exercise we also evaluated all the libs used by webgoat,
to verify if, and which of these libs may have any vulnerability.
For that, it was used OWASP Dependency-Check
(https://owasp.org/www-project-dependency-check/)
that takes the .jar file as input and generate a report of all the
used libs in the project.